+++ CHATGAME/Assets/Resources/image/UI/Panel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 34b21f255c3fa184b8d184f1cf3edc3f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
