+++ CHATGAME/Assets/Scripts/UIPanel/RecycleScrollView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 825064b0cb5718248b585bf915786084
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
