+++ CHATGAME/Assets/Scripts/UIPanel/SlotScroll.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d9fbd73fd3c93804eb8ef8c0635f8ac5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
